streamlit/requirements.txt
streamlit==1.32.0
pandas==2.2.1
SQLAlchemy==2.0.28
PyMySQL==1.1.0
python-dotenv==1.0.1
